alg: import single-link max-min fairness implementation
This is from a previous, unpublished project that needed a fast,
unweighted single-link max-min fairness implementation.

The plan is to use this as a goalpost for the B4 link waterfill
computation. It will probably remain much slower, but right now
there is no sense of whether it is 10x slower, 100x slower, or
1000x slower.